backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.6
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
pydantic==2.7.1
